God hates me
When we sincerely repent, the Lord promises that He will forgive us and remember our sins no more (D&C 58:42). Though repentance can be difficult and painful, the rewards of a clear conscience and spiritual healing are more than worth it. As we frequently repent, the Lord will help us become the men and women that He knows we can be.